feat: adds interceptor support to makeRequest
Adds support for response interceptors on the `makeRequest` utility. Response interceptors support a `onFulfilled` and `onRejected` callback which are processed like this:

- **response interceptor onFulfilled**: after processing the response from a successful `fetch` call with a non-error response into a response object and serializing the JSON body (if applicable)
- **response interceptor onRejected**: after processing the response from a successful `fetch` call with an error response into a response object and serializing the JSON body (if applicable)

Adds an axios-like interface to `RestClient` so the interceptors can be added on a per-client basis. Note that at the moment only a single response interceptor is supported. We might decide to switch to using `axios` in place of the `makeRequest` function so I didn’t want to go too far with adding logic here.

Changes the `makeRequest`’s list of positional parameters to a config object because that’s much easier to handle and makes our live writing tests easier.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>