chore: Remove notification manager (#1141)
Removes the NotificationManager, i.e. the little bell icon/button in the header, along with the modal that is opened when you click on that button.

This PR already removes quite surprising amount of associated code that was used to do this, but we can probably remove a little more as we no longer really need to load all the meshes at the beginning of the application to show this. (Even if we kept this we should probably have changed this to only make the call when the modal was opened i.e. lazily, when the associated modal becomes visible in the DOM)

Signed-off-by: John Cowen <john.cowen@konghq.com>