fix(charts): Make sure charts have some sort of loading state (#1161)
The code around loading our data for the overview charts has javascript loading states, but those states are **_not_** used to block the visibility/rendering of the rendered vue components.

You can see the javascript loading state here.

https://github.com/kumahq/kuma-gui/blob/39a4d9bb85f7ef29889622e0d29467acbbd7e635/src/app/main-overview/components/OverviewCharts.vue#L319

You'll notice that we don't actually then use this state in the Vue template at all.

In OSS this doesn't cause issues, but when trying elsewhere it does. It seems as though this, along with some sort of additional strange-ness in the Chart plugin we use that it can result in charts with no coloring.

Theoretically, this should all work due to Vue's reactive state but for some reason Vue's reactiveness doesn't seem to be working. Actually using the loading state in the template makes it work.

Signed-off-by: John Cowen <john.cowen@konghq.com>