chore: Removes last HTTP call to `/` and moves bootstrapping (#1157)
This PR began as removing the last HTTP API call to `/` that we have. Whilst doing this I noticing that we just invisibly failed on any errors in our 'bootstrapping' HTTP calls. I decided it was finally time to move our bootstrapping functionality into `App.vue` where we could catch any errors and show a proper error state.

---

### Removing the last HTTP API call to `/`

Previous to this PR we would make a call to `/` to see if that call was successful, if it wasn't successful we would just do nothing and miss out a bunch of HTTP calls we depend on further into the application, not great.

If we were successful, we wouldn't use any of the data from the API call, but instead make another bunch of API calls. Then if any of those failed, do nothing, and therefore be missing a bunch of data we then depend on further into the application, not great.

I don't quite understand why we would make a call to an endpoint just to see if its successful even though we don't use any of its data, when it would make more sense to use an endpoint that we actually use data from.

All the loading and error problems here led me to also...

### Move our 'bootstrapping' HTTP calls to App.vue

As we've been wanting to do this for quite a while, I figured if I did this now I could also use all our existing loading and error states to fail gracefully. 

---

### Lastly:

There is still so much stuff we can remove here now, but as its really intertwined throughout the app it takes quite a bit of picking apart. i.e. we still get a list of meshes at the start of the application, when we shouldn't need to; we still get a load of policy information at the start of the application when we shouldn't need to. There is ongoing work to change things so that we only load information for the page we are on directly from the backend every single time we load a page (as we want our data to be up-to-date and not stale), so hopefully we can pick this stuff out once thats there.

Our `bootstrap` function now does nothing in production, it is decorated with MSW functionality in development. Whilst this looks a little weird having an empty function, we can at some point noodle around with this as well, just I didn't want to do too much more here.

---

Signed-off-by: John Cowen <john.cowen@konghq.com>

Signed-off-by: John Cowen <john.cowen@konghq.com>