refactor: Cleans up RouteView code and adds basic docs (#1143)
The `RouteView`/`RouteTitle` and `AppView` components were initially made in a closed manner to PoC the different concept and then roll it out across the applications.

I'd always planned on coming back to tidy things up inside the components and add some documentation to help any external contributors or anyone unfamiliar with our codebase. This PR does that.

There was the initial concept of automatic navigation building/navigation selecting in `RouteView`, which we were using but pulled out and therefore we aren't using just yet. I decided to remove this entirely in this the clean up and add afresh at a later date. In the PR you will see things related to `children` and `module=""` being removed, this deletion is related to this.

Signed-off-by: John Cowen <john.cowen@konghq.com>