chore: Add title and actions slots to AppView (#1180)
Add title and actions slots to `AppView`

The vast majority of `AppView`s are going to be displayed with a title and probably a set of actions in either a dropdown or just a single button. This PR therefore adds title and actions slots to `AppView` so we can add additional markup/styling for that in one place.

I've used `section` for section content, which along with the `header`s and `h*`, will should give us more correct structuring/document outline.

`h*` tags are defined per route component so we can choose which level we need depending on the level of route/header nesting.

I've also added `render=true` to `RouteTitle` so you can use it within a `h*` to provide the header and the HTML title using a single element. Later we will probably default to true and turn it off for the rarer time when you don't want it to render. Right now I would rather gradually roll this out across the application and then switch in one go.

Lastly, I removed a `kcard-stack` CSS rule as we don't seem to be using it (see convo offline).

**Note: The change in the application code here is for PR example only, and it will be removed post approval before merging, and initial usage or this functionality will be done elsewhere.** If this approach seems fine I will also work on integration elsewhere and update the `AppView` docs.

---

Ok this grew a tiny bit bigger than I'd originally planned as I noticed we could reuse `AppView#title` and `AppView#actions` instead of the extra `WizardTitleBar` component. Less code, more reuse, less to learn.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>