chore: makes services mock more consistent
Updates the mocks surrounding the /service-insights and /external-services endpoints. They now generate resource names including their service type so we can use that in the mocks to associate external services to service insights with `serviceType: 'external'`. This allows us to show consistent data in the UI like not including fields that don’t make sense for external services, etc.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>