feat: implements new service detail view design (#1313)
* chore: makes services mock more consistent

Updates the mocks surrounding the /service-insights and /external-services endpoints. They now generate resource names including their service type so we can use that in the mocks to associate external services to service insights with `serviceType: 'external'`. This allows us to show consistent data in the UI like not including fields that don’t make sense for external services, etc.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* feat: implements new service detail view design

Implements the new service detail view design.

Fixes an issue with the service detail view always trying to load an external service even when the corresponding internal service is of service type "internal".

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore(e2e): randomizes default service type

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore(e2e): removes service type suffixes from service tests

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>