feat: improves error block component
Makes `error` a required prop as we already always provide it. As a result, some of the template could be simplified a bit.

Shows `error.detail` as title instead of the default "An error has occurred while trying to load this data." (which is now used as a fallback when `error.detail` isn’t set, i.e. when the error isn’t an `ApiError`).

Always shows `error.message` beneath the title.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>