refactor: loads policy type in components instead of globally (#1350)
* refactor: uses policy type instead of policy path in URLs

Replaces the route parameter `policyPath` with `policyType` for policy routes. This is much simpler to manage in the application because in many places, we already know the policy type name and don’t need anything else. Only in the policy views do we need to know the corresponding API path, but we can now isolate the mapping of policy type name to policy type path to those components. This in turn allows us to get rid of all store-kept policy type data and one app boot request.

Changes how navigating to the policy list view works to use an arbitrary (i.e. CircuitBreaker) policy type as the initially selected one instead of determining the first populated type on the route level.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor: switches back to using policy type API paths in URLs

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore: adds missing fields

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor: gives AppView title slot access to policy types

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* fix: loading and error block being in wrong order

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore: undoes some changes in DataplanePolicies

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* fix: loading block order

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore(e2e): moves DataplanePolicies test into Cypress

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>