feat(zones): moves configs into dedicated tabs
Moves the config shown in the Zone detail view into a new Config tab.

Adds a new YAML to the Zone Ingress/Egress detail views showing their respective configs. With this, I also changes the data source keys that fetch `ZoneIngressOverview`s and `ZoneEgressOverview`s to make clear when one is loading the non-Overview resource and when one is loading an Overview.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>