feat: improves error block (#1348)
* chore: adds trace ID to error title

Adds the trace ID from API errors to the error title so it’s visible when logging errors.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore: adds the trace ID to the standard error block

Adds the trace ID to the standard ErrorBlock component so it’s visible in the UI.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore: uses error block for all list views

Uses the standard ErrorBlock component for errors in list views instead of the built-in KTable error as it has more information (like the added trace ID).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* feat: improves error block component

Makes `error` a required prop as we already always provide it. As a result, some of the template could be simplified a bit.

Shows `error.detail` as title instead of the default "An error has occurred while trying to load this data." (which is now used as a fallback when `error.detail` isn’t set, i.e. when the error isn’t an `ApiError`).

Always shows `error.message` beneath the title.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>