chore: Removes store config module and AppErrorMessage (#1442)
Entirely removes the store config module plus the `AppErrorMessage`.

Our `bootstrap` function is currently empty in production but still gets called in production due to the fact that during development that is where we add our mocking. I'm hoping to move this soon, at which point the `bootstrap` function can be removed entirely.

I also removed `$.storeConfig` and mostly inlined that. The `store` has been gradually reducing in size and will hopefully soon be removed entirely.

Signed-off-by: John Cowen <john.cowen@konghq.com>