chore: Adds YAML undefined type to aid testing (#1452)
See https://github.com/kumahq/kuma-gui/pull/1316#discussion_r1293326627

We had a bit of discussion a while back on the inability to omit a property from a merge (i.e. mark it as undefined). This adds a `!!js/undefined` YAML tag to let you do that.

I don't think we'll ever need the ability to specify between omitted vs set to `undefined` but if we do we can just add a new tag to cope for that.

Signed-off-by: John Cowen <john.cowen@konghq.com>