chore(zones): tweaks create UI content formatting
Changes the base ul and ol padding so that it doesn’t overflow its content container narrowly.

Overrides the KCard body font size specifically for forms where we expect general body text to have the same font size (`$kui-font-size-40`).

Tweaks the gaps between various form elements and groups to make it more clear which elements belong together in some way or another.

Stops using `k-input-label` as a styling utility as it will be removed in a future version of Kongponents.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>