chore(e2e): Maintain history of requests instead of Cypress request spies (#1482)
Back when I started adding our Cypress support I spent a lot of time trying to get some sort of request assertion working in a nice way but didn't get there, but I wanted to come back and make it nicer at some point.

> After working with Cypress's intercept and request aliasing this PR I realized that request asserting ideally needs intercept to maintain a history of all requested API calls...
> This PR doesn't take this approach yet and will require a bit of reworking. The method I used here gets us to a point where we can write many more tests and come back to change this up. I want to add more tests and get this into CI before I come back to that though.

https://github.com/kumahq/kuma-gui/pull/828

Up until now occasionally you sometimes have to add extra steps to "tag" a request so Cypress can wait on it. Ideally, as we automatically mock out every single URL so you don't have to, we should just be able to say "tell me this request has happened" without have to tag it first. There were other little issues with this that I can't quite remember the details of.

This PR, instead of using Cypress, maintains a separate history of API requests, and then checks that history or waits on additional entries to that history in order to then perform an assertion as to whether it happened or not (timing out after 4000)

I did a few checks here first to see if the previous request asserting was causing the odd flake or difficulty and it looks like it was, you can see in the last lot of commits where I've added a quick "check for a DOM element first" and then remove that again once this was in. I also checked this was working by dropping a setTimeout in a `KumaApi` call before it made the actual HTTP request, and the new approach would wait until that request had happened before the step would move on, I also made the setTimeout really big to make sure it failed.

I've PRed this now even though I would like to rearrange some things here generally, just as I have in mind that this PR might be immediately beneficial as is, so if we get an approval here and merge I can work on moving things around slightly as we are behind the gherkin layer so my fiddling around in an upcoming PR won't affect anything.

Closes https://github.com/kumahq/kuma-gui/issues/1332

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>