fix: service dpp list linking to wrong views (#1641)
* fix(services): dpp list linking to wrong views

Fixes the service detail view’s DPP list linking to the wrong views depending. This change makes it so that the links in this list are always correct and independent from `props.gateways`.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* fix(services): check for gateway being inverted

Fixes the service detail view’s DPP list passing `props.gateways` as `true` when the first item **isn’t** a gateway and as `false` when the first item **is** a gateway. This view is still flawed, but it’s less broken now.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>