feat(app-collection): systematically add linked name and "Go to details" cells
Systematically adds a “name” cell to AppCollections. The name cell will always be linked as long as either a summary or detail route is provided.

Systematically adds a “Go to details” link to AppCollections that have a summary route. This is intended as the primary quick interaction to go from list to detail view when a summary view exists for that list view. This becomes relevant because the default row click interaction for list views with a summary view is to open the summary view (and no longer the detail view).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>