chore(meshes): avoids data source conflict
Avoids a potential data source conflict with the URI `/meshes` which was recently changed to resolve to a MeshInsight collection response. The existing data source for `/mesh-insights` has the same implementation and is better suited in terms of its name for views accessing `MeshInsight`s. For the OnboardingNotification component, the `/meshes` data source was changed to resolve to `Mesh`es again as it doesn’t need mesh insights. For the MainOverviewView component, the `/meshes` data source was changed to the `/mesh-insights` data source.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>