feat(control-planes): update onboarding alert logic
Changes the logic for showing the onboarding alert to rely on data from the `/global-insight` API and to only show the alert is there is exactly one mesh (we decided to drop the extra conditions for that mesh being called “default” and there being no dataplanes).

Persists the onboarding alert dismiss state in local storage.

Moves the onboarding alert code into the control planes module and adds the OnboardingAlert component as a component in its DI service configuration.

Adds a link to onboarding to the help menu.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>