feat(control-planes): update onboarding alert logic (#1746)
* chore(meshes): avoids data source conflict

Avoids a potential data source conflict with the URI `/meshes` which was recently changed to resolve to a MeshInsight collection response. The existing data source for `/mesh-insights` has the same implementation and is better suited in terms of its name for views accessing `MeshInsight`s. For the OnboardingNotification component, the `/meshes` data source was changed to resolve to `Mesh`es again as it doesn’t need mesh insights. For the MainOverviewView component, the `/meshes` data source was changed to the `/mesh-insights` data source.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* feat(control-planes): update onboarding alert logic

Changes the logic for showing the onboarding alert to rely on data from the `/global-insight` API and to only show the alert is there is exactly one mesh (we decided to drop the extra conditions for that mesh being called “default” and there being no dataplanes).

Persists the onboarding alert dismiss state in local storage.

Moves the onboarding alert code into the control planes module and adds the OnboardingAlert component as a component in its DI service configuration.

Adds a link to onboarding to the help menu.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* chore: moves onboarding alert into onboarding module

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor(onboarding): decorate ControlPlaneStatus with OnboardingAlert

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>