feat(app-collection): systematically add linked name and details cells
Systematically adds a “name” cell to AppCollections. The name cell will always be linked as long as either a summary or detail route is provided.

Systematically adds a “Go to details” link to AppCollections that have a detail route. This is intended as the primary quick interaction to go from list to detail view. This becomes important because the default row click interaction for list views with a summary view is to open the summary view (and no longer the detail view) making the interaction to go from list to detail view take two steps instead of one.

Integrates the actions dropdown into AppCollection because it’s a lot of unnecessary repetition. Users of AppCollection now pass dropdown items to AppCollection using the new optional `actions-items` slot (e.g. used by the Zone list view for the Delete item).

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>