chore: Add automatic typing to source internals (#1525)
This PR adds automatic types to the `"/services/:name": (params) => { params.name }` sources. The one little tiny downside is that currently you have to use `defineSources`, but you do maintain the key names (or the URIs) in the result type, which is important and why a simpler type definition wouldn't have been good enough for me.

Notes:

- We maintain the keys in the resulting definition, I don't want to lose these by doing something simpler like `Record<string,...>`
- `params` are now automatically inferred to strings based on the URI
- We don't assume things like `'all' | 'delegated' | 'builtin'` for a string that could come from the browser URL. i.e. `params.*` are always just strings not unions. I add a little runtime validation here to use a default if somehow something else ends up in the URL.
- `PaginationParams` are still typed as before, I'll potentially also make these strings, just not thought past that yet.

The plan is to eventually have a similar approach on The Other Side (i.e. the view layer), but both of these approaches should maintain some sort of separation similar to what we currently have, and should be as automatic as possible.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>