chore(dataplanes): fix rare flake
Fixes a rare flake that is caused by the mocked API endpoint for Dataplane details not removing the mTLS object. There was a negative assertion trying make sure that’s not the case, but it was executed before any content even rendered most of the times. When it didn’t _and_ the mock data rolled the dice for an expired cert, the flake happened.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>