feat(dataplanes): Add ServiceTraffic component (#1807)
This PR adds 2 things:

1. A `ServiceTraffic` component and related sub-components. This is the component only and is not hooked up to any data.
2. In order to be able to build/PR/merge the above component in isolation without adding it to our actual application. I've added a very rough sandbox in our docs that can be seen via `yarn docs:dev` (we can add PR previews at a later date which will be much better)

## ServiceTraffic component

I'm looking for feedback on the component itself, plus its subcomponents. Hopefully it's clear where the data will be plugged into this in a later PR and finally integrated into our application.

## Docs/Component Sandbox

I wanted to start getting this in as I would really like to use this more for building/manually testing/previewing our reusable components. It has the additional benefit of being able to work on, PR and merge components that aren't yet used in our application. Almost like feature flags but instead of feature flagging our in-production application, we just use a completely different application. In the future we should be able to gradually build entire features in isolation using this sandbox application, then just enable them in production when we want to.

As you can see the code here is very much in-progress. I went out of my way not to alter any of our application code just yet, but I think some injection related things will change ever so slightly to allow us to reuse things better.

Lastly I'd be happy to close this PR either way and wait for everything to be integrated, but I wanted to get feedback early on both of the above features and I also wanted to do a final CI check on the sandbox stuff and start thinking about CI integration. Let me know either way. I'd also be happy splitting these features into separate PRs, but really in order to test/experience one you need the other.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>