refactor(dataplanes): move warnings and cert expired utilities into data transformation file (#1835)
* chore(dataplanes): fix rare flake

Fixes a rare flake that is caused by the mocked API endpoint for Dataplane details not removing the mTLS object. There was a negative assertion trying make sure that’s not the case, but it was executed before any content even rendered most of the times. When it didn’t _and_ the mock data rolled the dice for an expired cert, the flake happened.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor(dataplanes): move warnings and cert expired utilities into data transformation file

Moves the warnings and cert expired functionality into the dataplanes data transformation file in the form of the new `getIsCertExpired` and `getWarnings` utilities.

Adds new `getDataplaneType` utility.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor(dataplanes): make all data transformers take a DataplaneOverview parameter

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* refactor(dataplanes): status utility edge case for gateways

Removes the `inbounds.length === 0` check for the Dataplanes status utility. It’s not needed because non-gateway Dataplanes always have an inbound.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>