chore(dataplanes): Add data source for traffic retrieval (#1839)
Adds a new `DataSource` `source` (in application code) for retrieving `stats` endpoint and parsing it into something we can use for the `DataPlaneTraffic` component.

The aim of this PR is to get something we can use in the application after this PR is merged, so I want to figure out what properties we should be using for the rendered stats here. For the moment I did what I'd understood was correct but then it didn't seem to show me anything, so I changed some things so that I had some numbers appearing. From what I understand this all depends on exactly what it is we want to show. We should be able to decide this and just tweak some words to show what we want.

Notes:

- I used an entirely different DataSource URI, this may or may not be merged into the one we currently use in the application, I kinda wanted to avoid touching application code for the moment.
- I commented out the stats in the title blocks (I'll note inline). I really want to get the names of the stats we want to be using here right before we finish these up.
- The mock stats I'm using are from an actual install with some requests thrown at it, so it should at least show me something other than zeros.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>