chore: Reduce line-height overwrites, add base font-size (#1816)
I keep noticing several bits of weirdness to do with line-heights and font-sizes. This PR deals mainly with line-heights and centralizes them using relative units in our base stylesheet.

The font-size addition was a late amend when I noticed some weirdness there also. At least for the moment, this amend means we can control our base font size, which is important for us seeing as we use rems quite a bit.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>