feat(dataplanes): Use service name for naming the inbounds instead of localhost (#1932)
- There is also a fix in here for showing the aggregate passthrough traffic correctly.
- Overall there is quite a lot of work here in beginning to flesh out the mock for the `stats` endpoint which includes making sure our stateless mocks use "fake state" by setting the seed used for randomization based on the stateful URL, therefore making sure the randomness produces the same results on two different API URLs.

Important note:

I've amended the rendering above slightly to what we want here so I'm taking this PR out of draft, but I have not yet changed the way we pick out the data we want (so I'm still overusing `find`s). I'm pretty sure I want to reshape our inbounds everywhere to be an object keyed by port instead of an array, that way we can just refer to the port of the inbound to find the thing rather than having to search through an array. I rather do that in a follow up though especially considering this feature is currently feature flagged, so I'm taking this out of draft for review.

Signed-off-by: John Cowen <john.cowen@konghq.com>