feat(dataplanes): add basic dataplane viz summary drawers (#1997)
Adds initial inbound/outbound summary drawers for the dataplane viz:

Mostly a copy paste from previous impl. of drawers apart from:

- Two way binding boolean route params no only execute once in our RouteView tree (only executing for the current route)
- `inbounds` and `outbounds` now have a `protocol` property derived from the tag but defaulting to `tcp`.

The contents of the drawers is purposefully very minimal until we know exactly what we want to show in those, I stopped short of filling the straightforwards things out as I wanted to talk about the usefulness of showing the exact same information in the drawer as we can already see in the viz itself (i.e. number of 2xx, 5xx requests etc). So this PR is partly technical but also to start off that conversation.

Related:

- #1959 
- #1960

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>
Signed-off-by: John Cowen <johncowen@users.noreply.github.com>
Co-authored-by: Philipp Rudloff <philipprudloff@fastmail.com>