feat: sync refreshing between data viz and drawer (#2070)
This PR ensures that clicking the summary drawer `[Refresh]` button in the dataplane viz will refresh the data both in the drawer and in the main viz.

There is a bit of refactoring work here as I noticed that, in the specific case where you have 2 of the exact same DataSources on the page _and_ you had a refresh button, clicking that button would not refresh. Pretty sure this is the first instance where we have this exact case. So whilst this DataSource syncing wasn't quite functioning correctly in this case, its not currently the source of a bug. But, it did prevent me from getting this working without having to do the work in this PR.

- `SharedPool` now works via a little state machine with distinct `creating`, `acquiring`, `releasing` and `destroying` states, which make this a lot clearer plus makes it control what happens when these states happen in one place rather than having to define entirely separate functions for each state.
- I added `cacheControl: no-cache` to DataSourcePool whilst I was here seeing as I was testing all this out it kinda made sense to add this now.
- Moved all data-plane `/stats` requesting to use the same DataSource. This means all requests for stats for dataplanes perform the same parsing logic even if they don't need it, but they also use the same cache. Seeing as you mostly always navigate to `DataplaneDetail > [Overview]` on your way to look at the `DataplaneDetail > [Stats]` tab, you will no longer see a loading spinner for the `DataplaneDetail > [Stats]` page. Consequently, even if the parsing did take JS a little longer millisecond-wise to do, the user will not see any difference. In fact, given we now use the cache, we provide a better experience for the user as a result of not needing to show a loading spinner.
- I did a little bit of cleanup whilst I was here, mainly moving some DataSource related exports to `@/app/application`.

---

At some point soon, I would like to do more work here to make the caching a little more generic i.e. we would pass a cache in that has its own way to `has`, `get` and `set` and we would move the `cacheControl` code into the cache implementation itself. I'd also then deal with the cachePrefix string in the injectable cache itself. Seeing as the cache is now injected, it makes it configurable per application. Lastly `has` and `get` themselves will be moved to use an `async` interface. This work won't happen immediately but probably pretty soon. I didn't do this extra cache related work here as it here as it requires some amends in other code that I didn't want to do just yet.



Signed-off-by: John Cowen <john.cowen@konghq.com>