chore(policies): fix flaky list view test
Add a content-based assertion in the middle of the "Clicking the link goes to the detail page and back again" so that the "navigate" + "assert URL" + "navigate" chain doesn’t trigger the missing route params error.

Remove the `key` attribute from `PolicyList`. It no longer seems needed to trigger a re-render when switching between policy types.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>