chore: remove/move upgrade component (#2125)
Moves the upgrade markup to `ApplicationShell` and removes the old `UpgradeCheck` component.

The old component was a single use component and no longer had any sort of logic/JS in it, i.e. it was essentially a glorified single-use partial. Seeing as `ApplicationShell` is also single-use it probably better to just add the necessary markup there.

Signed-off-by: John Cowen <john.cowen@konghq.com>