feat(dataplanes): support visualization for built-in gateways
Handle built-in gateways in the new Dataplane detail view display.

Changes the Dataplane stats data source to expect a `service` parameter to account for that part being dynamic in the case of gateways and provides `'localhost_'` for standard proxies as the value for this parameter. This allows us to maintain one data source definition for stats.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>